requirements.txt
pytest~=8.3.3
selene>=2.0.0rc9
selenium~=4.24.0
allure-pytest>=2.13.5
python-dotenv==1.0.1